+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - feat: add FlutterDeckCodeHighlight widget
 - feat: make the background configurable for the whole slide deck
 - feat: allow to hide slides in the slide deck
+- feat: add presentation progress indicator
 - fix: drawer toggle
 - docs: update the example app and README.md